items

+++ src/main/java/com/saliqdar/billingsoftware/service/impl/CategoryServiceImpl.java
@@ -1,9 +1,12 @@
 package com.saliqdar.billingsoftware.service.impl;
 
 import com.saliqdar.billingsoftware.entity.CategoryEntity;
+import com.saliqdar.billingsoftware.exception.InternalServerErrorException;
+import com.saliqdar.billingsoftware.exception.NotFoundException;
 import com.saliqdar.billingsoftware.io.CategoryRequest;
 import com.saliqdar.billingsoftware.io.CategoryResponse;
 import com.saliqdar.billingsoftware.repository.CategoryRepository;
+import com.saliqdar.billingsoftware.repository.ItemRepository;
 import com.saliqdar.billingsoftware.service.CategoryService;
 import com.saliqdar.billingsoftware.service.FileUploadService;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +23,7 @@ public class CategoryServiceImpl implements CategoryService {
 
    private  final FileUploadService fileUploadService;
     private final CategoryRepository categoryRepository;
+    private final ItemRepository itemRepository;
     @Override
     public CategoryResponse add(CategoryRequest request, MultipartFile file) {
         String imageUrl=fileUploadService.uploadFile(file);
@@ -40,16 +44,20 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryResponse readByCategoryId(String categoryId) {
         CategoryEntity entity=categoryRepository.findByCategoryId(categoryId)
-                .orElseThrow(() -> new RuntimeException("Category not found with id: " + categoryId));
+                .orElseThrow(() -> new NotFoundException("Category not found with id: " + categoryId));
         return convertToResponse(entity);
     }
 
     @Override
     public void delete(String categoryId) {
         CategoryEntity entity=categoryRepository.findByCategoryId(categoryId)
-                .orElseThrow(() -> new RuntimeException("Category not found with id: " + categoryId));
-        fileUploadService.deleteFile(entity.getImgUrl());
-        categoryRepository.delete(entity);
+                .orElseThrow(() -> new NotFoundException("Category not found with id: " + categoryId));
+       boolean isDeleted= fileUploadService.deleteFile(entity.getImgUrl());
+        if (isDeleted) {
+            categoryRepository.delete(entity);
+        }else  {
+            throw new InternalServerErrorException("Unable to delete category with id: " + categoryId);
+        }
     }
 
     private CategoryEntity convertToEntity(CategoryRequest request) {
@@ -62,11 +70,13 @@ public class CategoryServiceImpl implements CategoryService {
 
     }
     private CategoryResponse convertToResponse(CategoryEntity entity) {
+        Integer itemCount= itemRepository.countByCategoryId(entity.getId());
         return  CategoryResponse.builder()
                 .categoryId(entity.getCategoryId())
                 .name(entity.getName())
                 .description(entity.getDescription())
                 .imgUrl(entity.getImgUrl())
+                .itemCount(itemCount)
                 .bgColor(entity.getBgColor())
                 .createdAt(entity.getCreatedAt())
                 .updatedAt(entity.getUpdatedAt())
